// Servers/models/Role.ts
/**
 * Represents a role within the system.
 *
 * @interface Role
 * @property {number} id - The unique identifier for the role.
 * @property {string} name - The name of the role.
 * @property {string} description - A brief description of the role.
 */

export interface Role {
  id: number;
  name: string;
  description: string;
}